+++ packages/camera/camera/android/src/main/java/io/flutter/plugins/camera/media/MediaRecorderBuilder.java
@@ -21,6 +21,7 @@ public class MediaRecorderBuilder {
   private final MediaRecorderFactory recorderFactory;
 
   private boolean enableAudio;
+  private Integer audioEncoder;
   private int mediaOrientation;
 
   public MediaRecorderBuilder(
@@ -42,6 +43,11 @@ public class MediaRecorderBuilder {
     return this;
   }
 
+  public MediaRecorderBuilder setAudioEncoder(Integer audioEncoder) {
+    this.audioEncoder = audioEncoder;
+    return this;
+  }
+
   public MediaRecorderBuilder setMediaOrientation(int orientation) {
     this.mediaOrientation = orientation;
     return this;
@@ -56,7 +62,9 @@ public class MediaRecorderBuilder {
     mediaRecorder.setVideoSource(MediaRecorder.VideoSource.SURFACE);
     mediaRecorder.setOutputFormat(recordingProfile.fileFormat);
     if (enableAudio) {
-      mediaRecorder.setAudioEncoder(recordingProfile.audioCodec);
+      mediaRecorder.setAudioEncoder(audioEncoder != null ? audioEncoder : recordingProfile.audioCodec);
+      // TODO can we assume that bitrate and sampling rate from the CamcorderProfile will also be valid if we change
+      //      the encoder?
       mediaRecorder.setAudioEncodingBitRate(recordingProfile.audioBitRate);
       mediaRecorder.setAudioSamplingRate(recordingProfile.audioSampleRate);
     }