+++ packages/camera/camera/lib/camera.dart
@@ -16,4 +16,5 @@ export 'package:camera_platform_interface/camera_platform_interface.dart'
         FocusMode,
         ResolutionPreset,
         XFile,
+        AudioFormatGroup,
         ImageFormatGroup;