+++ packages/camera/camera_platform_interface/lib/src/types/types.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+export 'audio_format_group.dart';
 export 'camera_description.dart';
 export 'resolution_preset.dart';
 export 'camera_exception.dart';