+++ packages/camera/camera_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+* Upgrade `camera_platform_interface` to 0.2.0
+
 ## 0.2.0
 
 * Initial release, adapted from the Flutter [I/O Photobooth](https://photobooth.flutter.dev/) project.